convert to fuse-memfd.
Rather than build an internal block storage system, delegate file's
backing storage to anonymous memfd_create file descriptors.

Next steps would be to add an IPC layer. I'd imagine that fuse-memfd
would listen on a Unix Domain Socket for requested file names, and,
if fuse-memfd has such a file, & use usocket to write the fd of the
file back to the requesting process.

This process would now have shared access to the in-memory file that
fuse-memfd is also using to in-memory store the file's contents.